drf_project/requirements.txt
django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3

redis[hiredis]==5.0.8
